Penyesuaian Admin

+++ database/migrations/2025_11_30_162735_add_foto_to_barangs_table.php
@@ -10,17 +10,20 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up()
-{
-    Schema::table('barangs', function (Blueprint $table) {
-        $table->string('foto')->nullable()->after('stok');
-    });
-}
+    {
+        Schema::table('barangs', function (Blueprint $table) {
+            if (!Schema::hasColumn('barangs', 'foto')) {
+                $table->string('foto')->nullable()->after('stok');
+            }
+        });
+    }
 
-public function down()
-{
-    Schema::table('barangs', function (Blueprint $table) {
-        $table->dropColumn('foto');
-    });
-}
-
+    public function down()
+    {
+        Schema::table('barangs', function (Blueprint $table) {
+            if (Schema::hasColumn('barangs', 'foto')) {
+                $table->dropColumn('foto');
+            }
+        });
+    }
 };
